Move file_utils para o mesmo nível de sps
