+++ packtools/file_utils.py
@@ -118,7 +118,7 @@ def create_temp_file(filename, content=None, mode='w'):
 def copy_file(source, target):
     tmpdir = tempfile.mkdtemp()
     fullpath_target = os.path.join(tmpdir, target)
-    
+
     logger.info(f'Copying file {source} to {fullpath_target}')
     return shutil.copyfile(source, fullpath_target)
