+++ packtools/sps/sps_maker.py
@@ -1,6 +1,6 @@
 from packtools.sps.models import packages, sps_package
 from packtools.sps.libs import async_download, reqs
-from packtools.sps.utils import file_utils
+from packtools import file_utils
 from packtools.sps import exceptions
